requirements.txt
cookiecutter==2.6.0
crypto==1.4.1
cryptography==38.0.4
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
httplib2==0.20.4
itsdangerous==2.2.0
jsonpickle==4.1.1
jsonpointer==2.3
oauthlib==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-slugify==8.0.4
requests==2.28.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
websocket-client==1.2.3
Werkzeug==3.1.3
notion-client
notion
Flask-WTF
Flask-Mail
wtforms
uwsgi